+++ traceplot.h
@@ -30,7 +30,7 @@ class TracePlot : public Plot
 public:
 	TracePlot(std::shared_ptr<AbstractSampleSource> source);
 
-    void paintMid(QPainter &painter, QRect &rect, range_t<off_t> sampleRange);
+    void paintMid(QPainter &painter, QRect &rect, range_t<size_t> sampleRange);
     std::shared_ptr<AbstractSampleSource> source() { return sampleSource; };
 
 signals:
@@ -43,7 +43,7 @@ private:
 	QSet<QString> tasks;
 	const int tileWidth = 1000;
 
-	QPixmap getTile(off_t tileID, off_t sampleCount);
-	void drawTile(QString key, const QRect &rect, range_t<off_t> sampleRange);
-	void plotTrace(QPainter &painter, const QRect &rect, float *samples, off_t count, int step);
+	QPixmap getTile(size_t tileID, size_t sampleCount);
+	void drawTile(QString key, const QRect &rect, range_t<size_t> sampleRange);
+	void plotTrace(QPainter &painter, const QRect &rect, float *samples, size_t count, int step);
 };
