+++ tunertransform.cpp
@@ -26,7 +26,7 @@ TunerTransform::TunerTransform(std::shared_ptr<SampleSource<std::complex<float>>
 
 }
 
-void TunerTransform::work(void *input, void *output, int count, off_t sampleid)
+void TunerTransform::work(void *input, void *output, int count, size_t sampleid)
 {
     auto out = static_cast<std::complex<float>*>(output);
     std::unique_ptr<std::complex<float>[]> temp(new std::complex<float>[count]);
